膜のハッシュ関数を，リンク先をひとつ以上先を見ることができるように変更．

+++ slim/src/mhash.h
@@ -7,6 +7,6 @@
 
 #include "lmntal.h"
 
-int mhash(LmnMembrane *mem);
+unsigned long mhash(LmnMembrane *mem);
 
 #endif
